Extract `withMenuState` into its own module
