+++ components/src/NavBar/MobileMenu.js
@@ -13,6 +13,7 @@ import MenuItem from "./MenuItem";
 import SubMenuItemList from "./MenuDropdown/SubMenuItemList";
 import theme from "../theme";
 import { subPx } from "../Utils";
+import { withMenuState } from "./withMenuState";
 
 const isSubMenu = menuItem => (menuItem.subMenuItems);
 
@@ -105,45 +106,5 @@ MobileMenuBase.propTypes = {
 MobileMenuBase.defaultProps = {
   menuData: null,
 };
-
-class MenuState extends React.Component {
-  constructor() {
-    super();
-
-    this.state = {
-      isOpen: false,
-    };
-
-    this.handleOnClick = this.handleOnClick.bind(this);
-  }
-
-  handleOnClick() {
-    const { isOpen } = this.state;
-
-    this.setState({
-      isOpen: !isOpen,
-    });
-  }
-
-  render() {
-    const { isOpen } = this.state;
-    const { children: renderMenu } = this.props;
-
-    return renderMenu({
-      isOpen,
-      handleMenuToggle: this.handleOnClick,
-    });
-  }
-}
-
-const withMenuState = MenuComponentWithoutState => ({
-  ...props
-}) => (
-  <MenuState>
-    {
-      menuState => <MenuComponentWithoutState menuState={ menuState } { ...props } />
-    }
-  </MenuState>
-);
 
 export default withMenuState(MobileMenu);
